VMS-12987: Fix a bug in QnStorageDb::baseFileName() function: add extension
<==== Mercurial data ====>
Branch: vms

+++ vms/server/nx_vms_server/src/recorder/storage_db.h
@@ -101,7 +101,7 @@ private:
     void onVacuumFinished(bool success);
     bool readDbHeader() const;
     QStringList allDbFiles(const QString& basePath) const;
-    QString baseFileName(int seqId);
+    QString baseFileName(int64_t seqId);
 };
 
 typedef std::shared_ptr<QnStorageDb> QnStorageDbPtr;
